+++ examples/swash_render/src/main.rs
@@ -44,6 +44,9 @@ fn main() {
     // The display scale for HiDPI rendering
     let display_scale = 1.0;
 
+    // Whether to automatically align the output to pixel boundaries, to avoid blurry text.
+    let quantize = true;
+
     // The width for line wrapping
     let max_advance = Some(200.0 * display_scale);
 
@@ -84,7 +87,8 @@ fn main() {
             ..Default::default()
         };
 
-        let mut builder = layout_cx.tree_builder(&mut font_cx, display_scale, &root_style);
+        let mut builder =
+            layout_cx.tree_builder(&mut font_cx, display_scale, quantize, &root_style);
 
         builder.push_style_modification_span(&[bold_style]);
         builder.push_text(&text[0..5]);
@@ -130,7 +134,7 @@ fn main() {
         // ============
 
         // Creates a RangedBuilder
-        let mut builder = layout_cx.ranged_builder(&mut font_cx, &text, display_scale);
+        let mut builder = layout_cx.ranged_builder(&mut font_cx, &text, display_scale, quantize);
 
         // Set default text colour styles (set foreground text color)
         builder.push_default(brush_style);