+++ parley/src/tests/mod.rs
@@ -4,5 +4,6 @@
 mod test_basic;
 mod test_cursor;
 mod test_editor;
+mod test_lines;
 mod test_wrap;
 mod utils;